feature: add debug getProperties cache so that no information is lost when collapsing and reexpanding items
